% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_map.R
\name{corr_map}
\alias{corr_map}
\title{Correlation Map Plotting Function}
\usage{
corr_map(
  data,
  features,
  corr_method = "pearson",
  color_scheme = "blueorange",
  plot_width = 450,
  plot_height = 450,
  title = "Correlation Map"
)
}
\arguments{
\item{data}{A tibble.}

\item{features}{A character vector. It contains names of numerical feature in str for correlation map plotting. It should contain at least 2 features.}

\item{corr_method}{A character string. The method to calculate correlation between features. The default is "Pearson", two other supported methods are "'kendall' and 'spearman'.}

\item{color_scheme}{A character string. The method to calculate correlation between features. The default is "Pearson", two other supported methods are "'kendall' and 'spearman'.}

\item{plot_width}{An integer. The width of the plot}

\item{plot_height}{An integer. The heiht of the plot}

\item{title}{A character string. The title of the correlation map}
}
\value{
corr A ggplot object. Tue correlation map plot.
}
\description{
Plot a correlation map with the given dataframe object and a str vecotr of numerical features.
Users are allowed to set multiple arguments regarding the setting of the correlation plot
including correlation method, color schemes, plot width, height, and plot title.
}
\examples{
df <- data(mtcars)
corr_map <- corr_map(df, c("Horsepower", "Displacement", "Cylinders", "Acceleration"))

}
